+++ src/core/services/user/user-impl.service.ts
@@ -1,9 +1,8 @@
 import { AxiosResponse } from "axios";
 import { map, Observable } from "rxjs";
 
+import { HttpClient } from "@/core/libs/http/http-client";
 import { UserDomain } from "@/domain/user.domain";
-
-import { HttpClient } from "@/libs/http/http-client";
 
 import { UserService } from "./user.service";
