+++ src/core/services/user/user-mock.service.ts
@@ -1,8 +1,7 @@
 import { BehaviorSubject, Observable } from "rxjs";
 
+import { HttpClient } from "@/core/libs/http/http-client";
 import { UserDomain } from "@/domain/user.domain";
-
-import { HttpClient } from "@/libs/http/http-client";
 
 import { UserService } from "./user.service";
