+++ src/core/services/user/user.service.ts
@@ -1,8 +1,7 @@
 import { Observable } from "rxjs";
 
+import { HttpClient } from "@/core/libs/http/http-client";
 import { UserDomain } from "@/domain/user.domain";
-
-import { HttpClient } from "@/libs/http/http-client";
 
 interface IUserService {
   url: URL;
